refactor(badges): move BadgeRepository and NIP-58 models/parser into the package
Relocates badge_repository.dart, nip58_badge_models.dart, and
nip58_badge_parser.dart from mobile/lib/services/badges into the new package's
lib/src, switching internal imports to package: URIs. Drops the unreachable
private Nip58BadgeParser._() constructor so the static-only class is fully
covered. No behavior change (#3866).
