+++ mobile/packages/badge_repository/lib/src/nip58_badge_models.dart
