+++ mobile/packages/badge_repository/lib/src/nip58_badge_parser.dart
@@ -1,9 +1,7 @@
+import 'package:badge_repository/src/nip58_badge_models.dart';
 import 'package:nostr_sdk/nostr_sdk.dart';
-import 'package:openvine/services/badges/nip58_badge_models.dart';
 
 class Nip58BadgeParser {
-  const Nip58BadgeParser._();
-
   static Nip58BadgeDefinition? parseDefinition(Event event) {
     if (event.kind != EventKind.badgeDefinition) return null;
